+++ packages/internal/bridge/sdk/src/constants/bridges.ts
@@ -16,7 +16,7 @@ export const ZKEVM_DEVNET_CHAIN_ID = '13433';
 /**
  * @constant {string} ZKEVM_TESTNET_CHAIN_ID - The chain ID for the zkEVM testnet (EIP-155 compatible format).
  */
-export const ZKEVM_TESTNET_CHAIN_ID = '13372';
+export const ZKEVM_TESTNET_CHAIN_ID = '13472';
 /**
  * @constant {string} ZKEVM_MAINNET_CHAIN_ID - The chain ID for the zkEVM mainnet (EIP-155 compatible format).
  */